Finalized basic XHR.
XHR is working in standards compliant browsers.  Might be a little more
efficient as now we can directly set some of the DOM element's methods,
versus having to add a listener. (Not sure if IE supports this, yet, but
about to find out...).

Also added string.[append,prepend] for returning formatted items.